WIP Integrate quantization into finetune script
